# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sphere English LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules:
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a module for a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleRequest"
      responses:
        "201":
          description: Module created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{id}:
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Update module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModuleRequest"
      responses:
        "200":
          description: Updated module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"

    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Lessons
  /lessons:
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"

    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonRequest"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /lessons/{id}/complete:
    post:
      operationId: completeLesson
      tags: [lessons]
      summary: Mark a lesson as completed by student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResponse"

  # Live Classes
...
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get message conversations for current user
      responses:
        "200":
          description: Message conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{userId}:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Get message thread with a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message thread
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/{id}/read:
    patch:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark a message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Announcements
  /announcements:
    get:
      operationId: getAnnouncements
      tags: [announcements]
      summary: Get announcements for current user's courses
      parameters:
        - name: courseId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"

    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement (teacher/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementRequest"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  # Leaderboard
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # Auth
...
    Module:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: integer
        lessonsCount:
          type: integer
      required:
        - id
        - courseId
        - title
        - order
        - lessonsCount
...
    CreateModuleRequest:
      type: object
      properties:
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        order:
          type: ["integer", "null"]
      required:
        - courseId
        - title
    UpdateModuleRequest:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        order:
          type: ["integer", "null"]

    # Lessons
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [video, document, quiz, live, audio, text, image]
        content:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        order:
          type: integer
        isCompleted:
          type: ["boolean", "null"]
      required:
        - id
        - moduleId
        - title
        - type
        - order
    LessonDetail:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [video, document, quiz, live, audio, text, image]
        content:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        order:
          type: integer
        isCompleted:
          type: ["boolean", "null"]
        quiz:
          $ref: "#/components/schemas/QuizDetail"
      required:
        - id
        - moduleId
        - title
        - type
        - order
    CreateLessonRequest:
      type: object
      properties:
        moduleId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [video, document, quiz, live, audio, text, image]
        content:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        order:
          type: ["integer", "null"]
      required:
        - moduleId
        - title
        - type
    UpdateLessonRequest:
      type: object
      properties:
        title:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
          enum: [video, document, quiz, live, audio, text, image, null]
        content:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        order:
          type: ["integer", "null"]

    # Live Classes
...
    QuizDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        lessonId:
          type: ["integer", "null"]
        courseId:
          type: ["integer", "null"]
        timeLimit:
          type: ["integer", "null"]
        passingScore:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        totalPoints:
          type: integer
      required:
        - id
        - title
        - passingScore
        - questions
        - totalPoints
    Question:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        type:
          type: string
          enum: [multiple_choice, true_false, fill_blank, matching, ordering]
        question:
          type: string
        options:
          type: ["array", "null"]
          items:
            type: string
        correctAnswer:
          type: string
        points:
          type: integer
        order:
          type: integer
      required:
        - id
        - quizId
        - type
        - question
        - correctAnswer
        - points
        - order
...
    ProgressResponse:
      type: object
      properties:
        lessonId:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
        pointsEarned:
          type: integer
      required:
        - lessonId
        - completed
        - pointsEarned
...
    Conversation:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        userRole:
          type: string
        lastMessage:
          type: string
        lastMessageAt:
          type: string
        unreadCount:
          type: integer
      required:
        - userId
        - userName
        - userRole
        - lastMessage
        - lastMessageAt
        - unreadCount
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        receiverId:
          type: integer
        content:
          type: string
        sentAt:
          type: string
        isRead:
          type: boolean
      required:
        - id
        - senderId
        - senderName
        - receiverId
        - content
        - sentAt
        - isRead
    SendMessageRequest:
      type: object
      properties:
        receiverId:
          type: integer
        content:
          type: string
      required:
        - receiverId
        - content

    # Announcements
    Announcement:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: ["integer", "null"]
        courseTitle:
          type: ["string", "null"]
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - authorId
        - authorName
        - createdAt
    CreateAnnouncementRequest:
      type: object
      properties:
        courseId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
      required:
        - title
        - content

    # Leaderboard
...